UPDATE CHINH SUA VAI PHAN CONTROLLER SU DUNG TOKEN

+++ src/main/java/org/grp8/swp391/controller/FavoriteController.java
@@ -1,6 +1,8 @@
 package org.grp8.swp391.controller;
 
 
+import jakarta.servlet.http.HttpServletRequest;
+import org.grp8.swp391.config.JwtUtils;
 import org.grp8.swp391.entity.Favorite;
 import org.grp8.swp391.entity.Listing;
 import org.grp8.swp391.entity.User;
@@ -27,6 +29,9 @@ public class FavoriteController {
     @Autowired
     private ListingService listingService;
 
+    @Autowired
+    private JwtUtils jwtUtils;
+
 
     @GetMapping("/user/{userId}")
     public ResponseEntity<?> getFavoriteByUser(@PathVariable String userId){
@@ -39,29 +44,32 @@ public class FavoriteController {
     }
 
     @PostMapping("/create")
-    public ResponseEntity<?> addFavorite(@RequestParam String userId, @RequestParam String listingId) {
-        try {
-            User user = userService.findUserById(userId);
-            if (user == null) {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found");
+    public ResponseEntity<?> addFavorite(HttpServletRequest req, @RequestParam String listingId) {
+        try{
+            String token = jwtUtils.extractToken(req);
+            if (token == null || !jwtUtils.checkValidToken(token)) {
+                return ResponseEntity.status(401).body("Invalid or missing token");
             }
 
-            Listing listing = listingService.findById(listingId);
-            if (listing == null) {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Listing not found");
-            }
-
-            Favorite favorite = favoriteService.addFavorite(user, listing);
-            return ResponseEntity.status(HttpStatus.CREATED).body(favorite);
-
-        } catch (RuntimeException e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
+            String email = jwtUtils.getUsernameFromToken(token);
+            User user = userService.findUserById(email);
+            Listing lis = listingService.findById(listingId);
+            Favorite fav = favoriteService.addFavorite(user, lis);
+            return ResponseEntity.status(HttpStatus.CREATED).body(fav);
+        }catch(RuntimeException e){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
     @DeleteMapping("/remove")
-    public ResponseEntity<?> deleteFavorite(@RequestParam String userId, @RequestParam String listingId) {
+    public ResponseEntity<?> deleteFavorite(HttpServletRequest req, @RequestParam String listingId) {
         try {
-            User user = userService.findUserById(userId);
+            String token = jwtUtils.extractToken(req);
+            if (token == null || !jwtUtils.checkValidToken(token)) {
+                return ResponseEntity.status(401).body("Invalid or missing token");
+            }
+            String email = jwtUtils.getUsernameFromToken(token);
+            User user = userService.findUserById(email);
+
             Listing listing = listingService.findById(listingId);
             favoriteService.removeFavorite(user, listing);
             return ResponseEntity.ok("Removed from favorites successfully");
